refactor: replace hand-written TaskRow with type derived from DB schema
- Move src/database.types.ts → shared/database.types.ts so both the
  backend (src/) and frontend (frontend/src/) can import the generated
  Supabase types
- Define TaskRow in shared/types.ts using field types indexed from
  Row<"tasks">, eliminating the hand-written interface in
  frontend/src/types.ts
- Re-export TaskRow from frontend/src/types.ts for backward-compatible
  imports in frontend code
- Update all backend and test imports from src/database.types.js to
  shared/database.types.js
- Update CLAUDE.md to reflect the new file location and regeneration
  command

Column type changes in DB migrations are now caught at compile time
without needing to update two separate interface declarations.

Closes #592
